+++ src/main/kotlin/com/zepben/auth/server/Auth0AuthHandler.kt
@@ -1,18 +1,10 @@
-// Copyright 2019 Zeppelin Bend Pty Ltd
-// This file is part of zepben-auth.
-//
-// zepben-auth is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// zepben-auth is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with zepben-auth.  If not, see <https://www.gnu.org/licenses/>.
+/*
+ * Copyright 2023 Zeppelin Bend Pty Ltd
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
 
 
 package com.zepben.auth.server
@@ -25,19 +17,18 @@ import io.vertx.core.http.HttpHeaders
 import io.vertx.core.http.HttpMethod
 import io.vertx.core.json.JsonObject
 import io.vertx.ext.auth.User
+import io.vertx.ext.auth.authorization.RoleBasedAuthorization
+import io.vertx.ext.auth.authorization.WildcardPermissionBasedAuthorization
 import io.vertx.ext.web.RoutingContext
-import io.vertx.ext.web.handler.AuthHandler
-import io.vertx.ext.web.handler.impl.HttpStatusException
-import java.util.concurrent.atomic.AtomicBoolean
-import java.util.concurrent.atomic.AtomicInteger
-
+import io.vertx.ext.web.handler.AuthenticationHandler
+import io.vertx.ext.web.handler.HttpException
 
 class Auth0AuthHandler(
-    val authProvider: JWTAuthProvider,
+    private val authProvider: JWTAuthProvider,
     requiredClaims: Set<String>,
     private val skip: String? = null
 ) :
-    AuthHandler {
+    AuthenticationHandler {
 
     private val authorities = mutableSetOf<String>()
 
@@ -45,51 +36,31 @@ class Auth0AuthHandler(
         addAuthorities(requiredClaims)
     }
 
-    override fun addAuthority(authority: String): AuthHandler {
-        authorities.add(authority)
-        return this
-    }
-
-    override fun addAuthorities(authorities: Set<String>): AuthHandler {
+    private fun addAuthorities(authorities: Set<String>): Auth0AuthHandler {
         this.authorities.addAll(authorities)
         return this
     }
 
-    override fun authorize(user: User?, handler: Handler<AsyncResult<Void?>>) {
-        val requiredCount = authorities.size
-        if (requiredCount > 0) {
-            if (user == null) {
-                handler.handle(Future.failedFuture(HttpStatusException(403, "No user was found, you must authenticate first")))
-                return
-            }
-            val count = AtomicInteger()
-            val sentFailure = AtomicBoolean()
-            val authHandler =
-                Handler { res: AsyncResult<Boolean> ->
-                    if (res.succeeded()) {
-                        if (res.result()) {
-                            if (count.incrementAndGet() == requiredCount) {
-                                // Has all required authorities
-                                handler.handle(Future.succeededFuture())
-                            }
-                        } else {
-                            if (sentFailure.compareAndSet(false, true)) {
-                                handler.handle(Future.failedFuture(HttpStatusException(403, "Could not authorise all requested permissions. This is likely a bug.")))
-                            }
-                        }
-                    } else {
-                        handler.handle(Future.failedFuture(res.cause()))
-                    }
-                }
-            for (authority in authorities) {
-                if (!sentFailure.get()) {
-                    user.isAuthorized(authority, authHandler)
-                }
-            }
-        } else {
+    private fun authorize(user: User?, handler: Handler<AsyncResult<Void?>>) {
+        if (authorities.isEmpty()) {
             // No auth required
             handler.handle(Future.succeededFuture())
+            return
         }
+        if (user == null) {
+            handler.handle(Future.failedFuture(HttpException(403, "No user was found, you must authenticate first")))
+            return
+        }
+        for (authority in authorities) {
+            val authorization =
+                if (authority.startsWith("role:")) RoleBasedAuthorization.create(authority.substring(5))
+                else WildcardPermissionBasedAuthorization.create(authority)
+            if (!authorization.match(user)) {
+                handler.handle(Future.failedFuture(HttpException(403, "Could not authorise all requested permissions. This is likely a bug.")))
+                return
+            }
+        }
+        handler.handle(Future.succeededFuture())
     }
 
     override fun handle(ctx: RoutingContext) {
@@ -119,9 +90,7 @@ class Auth0AuthHandler(
             }
 
             // proceed to authN
-            authProvider.authenticate(
-                res.result()
-            ) { authN: AsyncResult<User> ->
+            authProvider.authenticate({ res.result() }) { authN: AsyncResult<User> ->
                 if (authN.succeeded()) {
                     val authenticated = authN.result()
                     ctx.setUser(authenticated)
@@ -130,10 +99,10 @@ class Auth0AuthHandler(
                     // proceed to AuthZ
                     authorizeUser(ctx, authenticated)
                 } else {
-                    if (authN.cause() is HttpStatusException) {
+                    if (authN.cause() is HttpException) {
                         processException(ctx, authN.cause())
                     } else {
-                        processException(ctx, HttpStatusException(401, authN.cause()))
+                        processException(ctx, HttpException(401, authN.cause()))
                     }
                 }
             }
@@ -142,7 +111,7 @@ class Auth0AuthHandler(
 
     private fun processException(ctx: RoutingContext, exception: Throwable?) {
         if (exception != null) {
-            if (exception is HttpStatusException) {
+            if (exception is HttpException) {
                 val statusCode = exception.statusCode
                 val payload = exception.payload
                 when (statusCode) {
@@ -184,14 +153,14 @@ class Auth0AuthHandler(
         // See: https://www.w3.org/TR/cors/#cross-origin-request-with-preflight-0
         // Preflight requests should not be subject to security due to the reason UAs will remove the Authorization header
         if (request.method() == HttpMethod.OPTIONS) {
-            // check if there is a access control request header
+            // check if there is an access control request header
             val accessControlRequestHeader =
                 ctx.request().getHeader(HttpHeaders.ACCESS_CONTROL_REQUEST_HEADERS)
             if (accessControlRequestHeader != null) {
                 // lookup for the Authorization header
                 for (ctrlReq in accessControlRequestHeader.split(",".toRegex()).toTypedArray()) {
                     if (ctrlReq.equals("Authorization", ignoreCase = true)) {
-                        // this request has auth in access control, so we can allow preflighs without authentication
+                        // this request has auth in access control, so we can allow preflights without authentication
                         ctx.next()
                         return true
                     }
@@ -209,7 +178,7 @@ class Auth0AuthHandler(
         val authorization = request.headers()[HttpHeaders.AUTHORIZATION] ?: run {
             handler.handle(
                 Future.failedFuture(
-                    HttpStatusException(401, "Missing Authorization header")
+                    HttpException(401, "Missing Authorization header")
                 )
             ); return
         }
@@ -217,11 +186,11 @@ class Auth0AuthHandler(
         try {
             val idx = authorization.indexOf(' ')
             if (idx <= 0) {
-                handler.handle(Future.failedFuture(HttpStatusException(400, "Badly formed Authorization header")))
+                handler.handle(Future.failedFuture(HttpException(400, "Badly formed Authorization header")))
                 return
             }
             if (authorization.substring(0, idx) != "Bearer") {
-                handler.handle(Future.failedFuture(HttpStatusException(401, "Missing Bearer token from Authorization header")))
+                handler.handle(Future.failedFuture(HttpException(401, "Missing Bearer token from Authorization header")))
                 return
             }
             handler.handle(Future.succeededFuture(authorization.substring(idx + 1)))
@@ -230,9 +199,9 @@ class Auth0AuthHandler(
         }
     }
 
-    override fun parseCredentials(context: RoutingContext?, handler: Handler<AsyncResult<JsonObject>>?) {
+    private fun parseCredentials(context: RoutingContext?, handler: Handler<AsyncResult<JsonObject>>?) {
 
-        if (skip != null && context!!.normalisedPath().startsWith(skip)) {
+        if (skip != null && context!!.normalizedPath().startsWith(skip)) {
             context.next()
             return
         }