+++ src/test/kotlin/com/zepben/auth/server/AuthConfigRouteTest.kt
@@ -1,18 +1,10 @@
-// Copyright 2019 Zeppelin Bend Pty Ltd
-// This file is part of zepben-auth.
-//
-// zepben-auth is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// zepben-auth is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with zepben-auth.  If not, see <https://www.gnu.org/licenses/>.
+/*
+ * Copyright 2023 Zeppelin Bend Pty Ltd
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
 
 
 package com.zepben.auth.server
@@ -48,7 +40,7 @@ class AuthConfigRouteTest {
     fun testHandle() {
         val expectedResponse: String = JsonObject().apply {
             put("authType", AuthMethod.AUTH0)
-            put("issuerDomain", "test-issuer")
+            put("issuer", "test-issuer")
             put("audience", "test-audience")
             put("tokenPath", "test-token-path")
             put("algorithm", "RS256")