+++ changelog.md
@@ -1,7 +1,9 @@
 # Evolve SDK connection library
 ## [0.8.0] - UNRELEASED
 ### Breaking Changes
-* None.
+* Use super pom version 0.34.x, which uses Vert.x version 4.4.6 (major version change 3 &rarr; 4).
+* Removed unused module `com.zepben.auth.server.vertx`
+* `AuthConfigRoute` now expects the key `issuer` instead of `issuerDomain` for the OAuth provider's hostname.
 
 ### New Features
 * None.